container_src/requirements.txt
bs4===0.0.1
jsonschema==4.18.4
python-dotenv==1.0
requests==2.31.0
telethon==1.28.5
psycopg2-binary==2.9.9
flask==3.1.2